refactor responsive collection extension to core
