+++ sample/Responsive/Startup.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
+// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
 using System;
@@ -33,9 +33,7 @@ namespace Sandbox
         public void ConfigureServices(IServiceCollection services)
         {
             // Add responsive services.
-            services.AddResponsive()
-                .AddViewSuffix()
-                .AddViewSubfolder();
+            services.AddResponsive();
 
             // Add framework services.
             services.AddMvc();
