+++ src/Detection.Core/src/Builders/DetectionCoreBuilder.cs
@@ -1,10 +1,10 @@
-﻿// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
+// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
 using System;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Wangkanai.Detection.Builder
+namespace Wangkanai.Detection.Builders
 {
     /// <summary>
     /// Helper functions for configuring detection core services.
