+++ src/Detection.Core/src/DetectionCoreCollectionExtensions.cs
@@ -5,7 +5,7 @@ using System;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Wangkanai.Detection;
-using Wangkanai.Detection.Builder;
+using Wangkanai.Detection.Builders;
 
 // ReSharper disable once CheckNamespace
 namespace Microsoft.Extensions.DependencyInjection
