+++ src/Responsive.Core/src/Abstractions/IResponsiveCoreBuilder.cs
@@ -1,11 +1,11 @@
-// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
+﻿// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Wangkanai.Responsive
 {
-    public interface IResponsiveBuilder
+    public interface IResponsiveCoreBuilder
     {
         IServiceCollection Services { get; }
     }
