+++ src/Responsive.Core/src/Exceptions/ResponsiveCoreBuilderArgumentNullException.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) 2019 Sarin Na Wangkanai, All Rights Reserved.
+// The Apache v2. See License.txt in the project root for license information.
+
+namespace System
+{
+    public class ResponsiveCoreBuilderArgumentNullException : ArgumentNullException
+    {
+        public ResponsiveCoreBuilderArgumentNullException(string paramName) : base(paramName) { }
+    }
+}
