+++ src/Responsive.Core/src/ViewBuilderExtensions.cs
@@ -10,21 +10,21 @@ namespace Microsoft.Extensions.DependencyInjection
 {
     public static class ViewBuilderExtensions
     {
-        public static IResponsiveBuilder AddViewSuffix(this IResponsiveBuilder builder)
+        public static IResponsiveCoreBuilder AddViewSuffix(this IResponsiveCoreBuilder builder)
         {
             if (builder == null) throw new ViewBuilderExtensionsSuffixArgumentNullException(nameof(builder));
 
             return AddViewLocation(builder, ResponsiveViewLocationFormat.Suffix);
         }
 
-        public static IResponsiveBuilder AddViewSubfolder(this IResponsiveBuilder builder)
+        public static IResponsiveCoreBuilder AddViewSubfolder(this IResponsiveCoreBuilder builder)
         {
             if (builder == null) throw new ViewBuilderExtensionsSubfolderArgumentNullException(nameof(builder));
 
             return AddViewLocation(builder, ResponsiveViewLocationFormat.Subfolder);
         }
 
-        private static IResponsiveBuilder AddViewLocation(this IResponsiveBuilder builder, ResponsiveViewLocationFormat format)
+        private static IResponsiveCoreBuilder AddViewLocation(this IResponsiveCoreBuilder builder, ResponsiveViewLocationFormat format)
         {
             builder.Services.Configure<RazorViewEngineOptions>(
                 options =>
